Added and refactored tests
